refactor(devicepolicy): merge channel halves into api.go, interface into settings_writer.go
Pure file consolidation, no behavior or API change:
- fetch.go + report.go -> api.go (and their tests -> api_test.go): the
  Fetcher and Reporter are the two halves of the same developer-mdm-agent
  HTTP channel and share config gating, auth, timeout, and error style.
- writer.go -> settings_writer.go: the Writer interface now lives beside
  its only production implementation.

go doc -all diff before/after confirms the exported surface is identical
apart from the reworded file references in comments.
